simple-mcp-server-test/requirements.txt
mcp==1.26.0
pandas==3.0.0
pyyaml
